rework waitWhileBusy
Replace Thread yielding with a blocking wait to improve cpu usage. As
engineBusy and engineMode are not final it is dangerous to synchronize
on them, therefore a ReentrantLock is used. For the blocking wait a
Condition is used that is signaled each time when the result of isBusy
might change.