SignaturePlugin: include note in message if update was not successful
In case of `extend U with x do R` the value `x` will not be a member of `U` (as the UpdateSet wasn't consistent `x` wasn't added to `U`).

Therefore any writes to a function with a signature `.. * U * .. -> ..` or `.. -> U` will report `x` not being a member of `U`,
although there is nothing wrong with such an update.