Engine: fix assertion error caused by emError
The main loop assumes that the engine is busy, however the handling of
the emError state sets engineBusy to false. This causes an assertion,
that was added in c750929c via #26, to trigger, which leads in debug
runs (-ea flag set) to a crashed Engine and a blocked TestEngineDriver.

Additionaly the engineBusy flag is set to false even when the Engine
is actually busy as the emError state is left to either emIdle or
emTerminating.

This commit 1) changes the assertion to not throw when in emError state
and 2) corrects the setting of engineBusy to be false only when no
ecTerminate / ecRecover command was handled.

TODO: This basic fix leaves the Engine in a Thread.yield loop until a
ecTerminate / ecRecover command is present. The polling of the
commandQueue should be replaced with a blocking wait.